+++ fmpz_vec/scalar_submul_si.c
@@ -25,5 +25,5 @@ _fmpz_vec_scalar_submul_si(fmpz * vec1, const fmpz * vec2, slong len2, slong c)
             fmpz_submul_ui(vec1 + i, vec2 + i, c);
     else
         for (i = 0; i < len2; i++)
-            fmpz_addmul_ui(vec1 + i, vec2 + i, -c);
+            fmpz_addmul_ui(vec1 + i, vec2 + i, -((ulong) c));
 }
